backend/requirements.txt
fastapi
gspread
pandas
sqlalchemy
psycopg2-binary
uvicorn
python-dotenv
httpx
